Bike Rental Prediction
Build regression models that will predict the total number of bike rentals in an hour for the dataset bikeRentalHourlyTrain.csv. This dataset includes a header row. Build models using neural networks, linear regression (you may use regularization techniques – lasso and ridge), and KNN.

Please note that in this dataset, cnt is the sum of casual and registered, so you will not want to include these attributes in your model. There are other columns (like the first one, which is just an identifier), which you will also not include. Your script will read in the entire file, but only use relevant columns in the model. Information about the columns is contained in the file bikeSharingDataset.names.

Again, you may want to experiment to tune parameters/architectures etc. Your final script should read in the bikeRentalHourlyTrain.csv file, create and train the models (one for each technique), print out the training MSE, and then read in the bikeRentalHourlyTest.csv file, predict your results, and output the test MSE (for each of the three techniques). Please use 5-fold cross validation.